Introduce wrappers for timers
This allows us to mock timers in tests and run them much faster.
